requirements.txt
numpy>=1.19.0
torch>=1.7.0
pybullet>=3.0.0
gym>=0.17.0
pyyaml>=5.3.1
matplotlib>=3.3.0
tqdm>=4.46.0
casadi>=3.5.5
do-mpc>=4.1.1